add micromark
I chose micromark because it is sync by default. Remark is async, which adds complexity to react components. However, there is a react-remark plugin which provides a remarkSync method which works just as well, and returns actual React nodes rather than requiring a dangerouslySetInnerHTML function. Both work just as well.