add async rehype cf plugin
This can be called as part of an async unified() function chain, and will attempt to retrieve additional metadata from Contentful directly to reduce layout shift on embedded markdown images. Couldn't figure out how to get Astro to render async RSC so this has to be done in an .astro component for now.